+++ Sources/KeyboardKit/_Essentials/KeyboardView.swift
@@ -197,8 +197,9 @@ public struct KeyboardView<
             keyboardView
         }
         .autocorrectionDisabled(with: autocompleteContext)
-        .opacity(shouldShowEmojiKeyboard ? 0 : 1)
+        .opacity(shouldShowKeyboard ? 1 : 0)
         .overlay(emojiKeyboard, alignment: .bottom)
+        .overlay(numberPad, alignment: .bottom)
         .foregroundColor(styleService.foregroundColor)
         .background(renderBackground ? styleService.backgroundStyle : nil)
         .keyboardCalloutContainer(
@@ -231,10 +232,11 @@ private extension KeyboardView {
         )
     }
     
-    var shouldShowEmojiKeyboard: Bool {
+    var shouldShowKeyboard: Bool {
         switch keyboardContext.keyboardType {
-        case .emojis: true
-        default: false
+        case .emojis: false
+        case .numberPad: false
+        default: true
         }
     }
 }
@@ -268,7 +270,6 @@ private extension KeyboardView {
     var emojiKeyboard: some View {
         emojiKeyboardContent
             .id(keyboardContext.interfaceOrientation)           // TODO: Temp orientation fix
-            .opacity(shouldShowEmojiKeyboard ? 1 : 0)
     }
 
     @ViewBuilder
@@ -285,6 +286,18 @@ private extension KeyboardView {
             ))
         } else {
             Color.clear
+        }
+    }
+
+    @ViewBuilder
+    var numberPad: some View {
+        if keyboardContext.keyboardType == .numberPad {
+            Keyboard.NumberPad(
+                actionHandler: actionHandler,
+                styleService: styleService,
+                keyboardContext: keyboardContext
+            )
+            .padding(10)
         }
     }